Refactor TopBar to support mobile and desktop styles
