+++ package.json
@@ -23,6 +23,7 @@
     "mobx-react": "^6.1.1",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
+    "react-media": "^1.9.2",
     "react-onclickoutside": "^6.8.0",
     "react-scripts": "3.0.1",
     "styled-components": "^4.2.0",
